Pull from main

+++ APIs/Application/ViewModel/SubcriptionModel/CreateSubcriptionModel.cs
@@ -10,6 +10,6 @@ namespace Application.ViewModel.SubcriptionModel
     {
         public long Price { get; set; }
         public string SubcriptionType { get; set; }
-        public float ExpiryDay { get; set; }
+        public int ExpiryDay { get; set; }
     }
 }
